requirements.txt
flask
cryptography
qrcode
pyzbar
pillow